Add <alt> atribute to salad.jpg on index.html
Forgot to add alt atribute to the end of <img src="./images/salad.jpg">.
Did not effect code, but if missing would make it hard for a user with
assistive technology to be able to have the image described to them.
Added it to the end of the img element.